requirements.txt
abstract-rendering==0.5.1
alabaster==0.7.3
appdirs==1.4.0
appscript==1.0.1
argcomplete==0.8.9
astropy==1.0.3
aws-shell==0.1.1
awscli==1.6.2
Babel==1.3
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
bcolz==0.9.0
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
binstar==0.11.0
bitarray==0.8.1
blaze==0.8.0
blockchain==1.3.3
blz==0.6.2
bokeh==0.9.0
boto==2.33.0
boto3==1.4.0
botocore==0.73.0
Bottleneck==1.0.0
bottlenose==1.0.1
cdecimal==2.3
certifi==2015.4.28
cffi==1.1.0
chainer==1.8.2
clyent==0.3.4
colorama==0.2.5
conda==4.0.8
conda-build==1.14.1
conda-env==2.4.5
configobj==5.0.6
cryptography==0.9.1
cssselect==0.9.2
Cython==0.22.1
cytoolz==0.7.3
datashape==0.4.5
decorator==4.0.9
docutils==0.12
ebaysdk==2.1.4
enable==4.5.1
enum34==1.0.4
eventlet==0.19.0
fastcache==1.0.2
feedparser==5.2.1
filelock==2.0.6
Flask==0.10.1
funcsigs==0.4
futures==3.0.5
fuzzywuzzy==0.12.0
gevent==1.0.1
gevent-websocket==0.9.3
google-api-python-client==1.5.3
graphcanvas==4.0.2
greenlet==0.4.10
grin==1.2.1
gunicorn==19.6.0
h5py==2.5.0
httplib2==0.9.2
idna==2.0
ipaddress==1.0.7
ipython==3.2.0
itsdangerous==0.24
jdcal==1.0
jedi==0.8.1
Jinja2==2.7.3
jmespath==0.5.0
jsonpickle==0.9.3
jsonschema==2.4.0
llvmlite==0.5.0
lxml==3.6.4
Mako==1.0.4
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.5.1
mock==1.0.1
multipledispatch==0.4.8
networkx==1.9.1
nltk==3.0.3
nose==1.3.7
numba==0.19.1
numexpr==2.4.3
numpy==1.11.1
oauth2client==3.0.0
odo==0.3.2
openpyxl==1.8.5
pandas==0.18.1
patsy==0.3.0
pep8==1.6.2
Pillow==3.3.1
ply==3.6
prettytable==0.7.2
prompt-toolkit==1.0.7
protobuf==3.0.0b2
psutil==2.2.1
psycopg2==2.6.2
ptyprocess==0.4
py==1.4.27
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyAudio==0.2.7
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyface==4.5.2
pyflakes==0.9.2
Pygments==2.1.3
pymongo==3.2.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
PySocks==1.5.7
pytest==2.7.1
python-amazon-simple-product-api==2.1.0
python-dateutil==2.5.3
python-Levenshtein==0.12.0
pytools==2016.2.1
pytz==2016.4
PyYAML==3.11
pyzmq==14.7.0
redis==2.10.5
requests==2.11.1
requests-cache==0.4.12
rope==0.9.4
rsa==3.4.2
ruamel-yaml===-VERSION
runipy==0.1.3
s3transfer==0.1.3
scikit-image==0.11.3
scikit-learn==0.17.1
scipy==0.18.0
SFrame==2.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spyder==2.3.5.2
SQLAlchemy==1.0.5
statsmodels==0.6.1
sympy==0.7.6
tables==3.2.0
tensorflow==0.9.0
terminado==0.5
toolz==0.7.2
tornado==4.3
tqdm==4.8.4
traits==4.5.0
traitsui==4.5.1
ujson==1.33
unicodecsv==0.9.4
update==0.4.4
uritemplate==0.6
urllib3==1.16
virtualenv==15.0.1
wcwidth==0.1.7
websocket-client==0.37.0
Werkzeug==0.11.11
xlrd==0.9.3
XlsxWriter==0.7.3
xlwings==0.3.5
xlwt==1.0.0
youtube-dl==2016.9.27
